Separate Stan code to a stats module.
